Address feedback and move back some classes
