+++ analytics-sidecar-client/src/testFixtures/java/org/apache/cassandra/sidecar/client/request/RequestExecutorTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.sidecar.analyticsclient.request;
+package org.apache.cassandra.sidecar.client.request;
 
 import java.util.Arrays;
 import java.util.List;
