+++ analytics-sidecar-client/src/testFixtures/java/org/apache/cassandra/sidecar/client/request/TimeSkewRequestTestParameters.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.sidecar.analyticsclient.request;
+package org.apache.cassandra.sidecar.client.request;
 
 import org.apache.cassandra.sidecar.client.RequestContext;
 import org.apache.cassandra.sidecar.common.ApiEndpointsV1;
