+++ analytics-sidecar-vertx-client/src/test/java/org/apache/cassandra/sidecar/analyticsclient/request/BaseVertxRequestTest.java
@@ -23,6 +23,7 @@ import org.apache.cassandra.sidecar.client.HttpClientConfig;
 import org.apache.cassandra.sidecar.client.RequestExecutor;
 import org.apache.cassandra.sidecar.analyticsclient.VertxHttpClient;
 import org.apache.cassandra.sidecar.analyticsclient.VertxRequestExecutor;
+import org.apache.cassandra.sidecar.client.request.RequestExecutorTest;
 import org.apache.cassandra.sidecar.client.retry.BasicRetryPolicy;
 import org.apache.cassandra.sidecar.client.retry.RetryPolicy;
